use block fetcher from combined client to fetch blocks by root
